chore: Moves figlet test in lib, and use mockito in cli
