+++ gradle/libs.versions.toml
@@ -9,8 +9,9 @@
 [versions]
 asciidoctorj = "2.5.10"
 asciidoctorjDiagram = "2.2.11"
-junit-jupiter = "5.10.0"
 assertj = "3.24.2"
+junit-jupiter = "5.10.0"
+mockito = "5.5.0"
 picocli = "4.7.5"
 
 [libraries]
@@ -25,6 +26,7 @@ assertj = { module = "org.assertj:assertj-core", version.ref = "assertj" }
 junit-jupiter-api = { module = "org.junit.jupiter:junit-jupiter-api", version.ref = "junit-jupiter" }
 junit-jupiter-engine = { module = "org.junit.jupiter:junit-jupiter-engine", version.ref = "junit-jupiter" }
 junit-jupiter-params = { module = "org.junit.jupiter:junit-jupiter-params", version.ref = "junit-jupiter" }
+mockito-core = { module = "org.mockito:mockito-core", version.ref = "mockito" }
 
 jetbrains-annotations = { module = "org.jetbrains:annotations", version = "24.0.1" }
