+++ jufmt-cli/src/test/java/io/github/bric3/jufmt/app/JufmtFigletCommandTest.java
@@ -11,165 +11,70 @@
 package io.github.bric3.jufmt.app;
 
 import io.github.bric3.jufmt.EmbeddedFigletFonts;
+import io.github.bric3.jufmt.Figlet;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
+import org.mockito.Mockito;
 
-import java.util.stream.Stream;
+import java.util.Arrays;
+import java.util.regex.Pattern;
 
 import static io.github.bric3.jufmt.app.JufmtTestUtil.jufmt;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.params.provider.Arguments.arguments;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 
 public class JufmtFigletCommandTest {
     @Test
-    @Disabled("No more default font, only random")
-    void choose_same_default_font() {
-        var result = jufmt("figlet", "bric3");
+    void no_option_means_random() {
+        try (var mockedStatic = Mockito.mockStatic(Figlet.class)) {
+            mockedStatic.when(() -> Figlet.render(eq("bric3")))
+                    .thenReturn("A rendered string by figlet");
 
-        assertThat(result.out())
-                .isEqualToNormalizingNewlines("""
-                             _          _      _____\s
-                            | |__  _ __(_) ___|___ /\s
-                            | '_ \\| '__| |/ __| |_ \\\s
-                            | |_) | |  | | (__ ___) |
-                            |_.__/|_|  |_|\\___|____/\s
-                                                    \s
-                           """);
+            var result = jufmt("figlet", "bric3");
+
+            assertThat(result.out()).isEqualToIgnoringNewLines("A rendered string by figlet");
+        }
     }
 
     @Test
     void can_pick_random_font() {
-        var result = jufmt("figlet", "-r", "bric3");
+        try (var mockedStatic = Mockito.mockStatic(Figlet.class)) {
+            mockedStatic.when(() -> Figlet.render(eq("bric3")))
+                    .thenReturn("A rendered string by figlet");
 
-        assertThat(result.out()).isNotBlank();
+            var result = jufmt("figlet", "--random", "bric3");
+
+            assertThat(result.out()).isEqualToIgnoringNewLines("A rendered string by figlet");
+        }
     }
 
-    @ParameterizedTest
-    @MethodSource("figletArguments")
-    void supports_figlet(EmbeddedFigletFonts font, String input, String expected) {
-        var result = jufmt("figlet", "-f", font.name(), input);
+    @Test
+    void can_render_all_font() {
+        try (var mockedStatic = Mockito.mockStatic(Figlet.class)) {
+            mockedStatic.when(() -> Figlet.render(eq("bric3"), any(EmbeddedFigletFonts.class)))
+                    .thenReturn("A rendered string by figlet");
 
-        assertThat(result.out())
-                .describedAs(font.toString())
-                .isEqualToNormalizingNewlines(expected);
+            var result = jufmt("figlet", "--all", "bric3");
+
+            assertThat(result.out()).contains(Arrays.stream(EmbeddedFigletFonts.values()).map(Enum::name).toList());
+            assertThat(
+                    Pattern.compile("A rendered string by figlet")
+                            .matcher(result.out())
+                            .results().count()
+            ).isEqualTo(EmbeddedFigletFonts.values().length);
+        }
     }
 
-    private static Stream<Arguments> figletArguments() {
-        return Stream.of(
-                arguments(EmbeddedFigletFonts.Sub_Zero,
-                          "jufmt",
-                          """
-                             __     __  __     ______   __    __     ______ \s
-                            /\\ \\   /\\ \\/\\ \\   /\\  ___\\ /\\ "-./  \\   /\\__  _\\\s
-                           _\\_\\ \\  \\ \\ \\_\\ \\  \\ \\  __\\ \\ \\ \\-./\\ \\  \\/_/\\ \\/\s
-                          /\\_____\\  \\ \\_____\\  \\ \\_\\    \\ \\_\\ \\ \\_\\    \\ \\_\\\s
-                          \\/_____/   \\/_____/   \\/_/     \\/_/  \\/_/     \\/_/\s
-                                                                            \s
-                          """),
+    @Test
+    void supports_passing_font() {
+        try (var mockedStatic = Mockito.mockStatic(Figlet.class)) {
+            mockedStatic.when(() -> Figlet.render("jufmt", EmbeddedFigletFonts._3d))
+                    .thenReturn("A rendered string by figlet");
 
-                arguments(EmbeddedFigletFonts.Lean,
-                          "jufmt",
-                          """
-                                                                                   \s
-                                  _/                _/_/                    _/     \s
-                                     _/    _/    _/      _/_/_/  _/_/    _/_/_/_/  \s
-                                _/  _/    _/  _/_/_/_/  _/    _/    _/    _/       \s
-                               _/  _/    _/    _/      _/    _/    _/    _/        \s
-                              _/    _/_/_/    _/      _/    _/    _/      _/_/     \s
-                             _/                                                    \s
-                          _/                                                       \s
-                          """),
+            var result = jufmt("figlet", "--font", EmbeddedFigletFonts._3d.name(), "jufmt");
 
-                arguments(EmbeddedFigletFonts.Elite,
-                          "jufmt",
-                          """
-                           ▐▄▄▄▄• ▄▌·▄▄▄• ▌ ▄ ·. ▄▄▄▄▄
-                            ·███▪██▌▐▄▄··██ ▐███▪•██ \s
-                          ▪▄ ███▌▐█▌██▪ ▐█ ▌▐▌▐█· ▐█.▪
-                          ▐▌▐█▌▐█▄█▌██▌.██ ██▌▐█▌ ▐█▌·
-                           ▀▀▀• ▀▀▀ ▀▀▀ ▀▀  █▪▀▀▀ ▀▀▀\s
-                          """),
-
-                arguments(EmbeddedFigletFonts.Calvin_S,
-                          "jufmt",
-                          """
-                           ┬┬ ┬┌─┐┌┬┐┌┬┐
-                           ││ │├┤ │││ │\s
-                          └┘└─┘└  ┴ ┴ ┴\s
-                          """),
-
-                arguments(EmbeddedFigletFonts.Bear,
-                          "jufmt",
-                          """
-                             _     _      _     _      _     _      _     _      _     _  \s
-                            (c).-.(c)    (c).-.(c)    (c).-.(c)    (c).-.(c)    (c).-.(c) \s
-                             / ._. \\      / ._. \\      / ._. \\      / ._. \\      / ._. \\  \s
-                           __\\( Y )/__  __\\( Y )/__  __\\( Y )/__  __\\( Y )/__  __\\( Y )/__\s
-                          (_.-/'-'\\-._)(_.-/'-'\\-._)(_.-/'-'\\-._)(_.-/'-'\\-._)(_.-/'-'\\-._)
-                             || J ||      || U ||      || F ||      || M ||      || T ||  \s
-                           _.' `-' '._  _.' `-' '._  _.' `-' '._  _.' `-' '._  _.' `-' '._\s
-                          (.-./`-'\\.-.)(.-./`-'\\.-.)(.-./`-'\\.-.)(.-./`-'\\.-.)(.-./`-'\\.-.)
-                           `-'     `-'  `-'     `-'  `-'     `-'  `-'     `-'  `-'     `-'\s
-                          """),
-
-                arguments(EmbeddedFigletFonts.smpoison,
-                          "jufmt",
-                          """
-                                                                       \s
-                              @@@ @@@  @@@ @@@@@@@@ @@@@@@@@@@  @@@@@@@\s
-                              @@! @@!  @@@ @@!      @@! @@! @@!   @!!  \s
-                              !!@ @!@  !@! @!!!:!   @!! !!@ @!@   @!!  \s
-                          .  .!!  !!:  !!! !!:      !!:     !!:   !!:  \s
-                          ::.::    :.:: :   :        :      :      :   \s
-                                                                       \s
-                          """),
-
-                arguments(EmbeddedFigletFonts.future,
-                          "jufmt",
-                          """
-                           ┏┓╻ ╻┏━╸┏┳┓╺┳╸
-                            ┃┃ ┃┣╸ ┃┃┃ ┃\s
-                          ┗━┛┗━┛╹  ╹ ╹ ╹\s
-                          """),
-
-                arguments(EmbeddedFigletFonts.pagga,
-                          "jufmt",
-                          """
-                          ░▀▀█░█░█░█▀▀░█▄█░▀█▀
-                          ░░░█░█░█░█▀▀░█░█░░█░
-                          ░▀▀░░▀▀▀░▀░░░▀░▀░░▀░
-                          """),
-
-                arguments(EmbeddedFigletFonts.bigmono9,
-                          "jufmt",
-                          """
-                                                            \s
-                                                            \s
-                              █            ▒██              \s
-                                           █░            █  \s
-                                           █             █  \s
-                            ███   █   █  █████  ██▓█▓  █████\s
-                              █   █   █    █    █▒█▒█    █  \s
-                              █   █   █    █    █ █ █    █  \s
-                              █   █   █    █    █ █ █    █  \s
-                              █   █   █    █    █ █ █    █  \s
-                              █   █▒ ▓█    █    █ █ █    █░ \s
-                              █   ▒██▒█    █    █ █ █    ▒██\s
-                              █                             \s
-                             ▒█                             \s
-                            ██▒                             \s
-                          """),
-
-                arguments(EmbeddedFigletFonts.rusto,
-                          "jufmt",
-                          """
-                            ┬┬ ┐┬─┐┌┌┐┌┐┐
-                          ┌ ││ │├─ │││ │\s
-                          └─┆┆─┘┆  ┘ ┆ ┆\s
-                          """)
-                        );
+            assertThat(result.out()).isEqualToIgnoringNewLines("A rendered string by figlet");
+        }
     }
 }
